# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Quote": {
				"properties": {
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"open": {
						"type": "number",
						"format": "double"
					},
					"high": {
						"type": "number",
						"format": "double"
					},
					"low": {
						"type": "number",
						"format": "double"
					},
					"close": {
						"type": "number",
						"format": "double"
					},
					"adjClose": {
						"type": "number",
						"format": "double"
					},
					"volumn": {
						"type": "number",
						"format": "double"
					},
					"symbol": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"open",
					"high",
					"low",
					"close",
					"adjClose",
					"volumn",
					"symbol"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExchangeRateResponseModel": {
				"properties": {
					"quotes": {
						"items": {
							"$ref": "#/components/schemas/Quote"
						},
						"type": "array"
					}
				},
				"required": [
					"quotes"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "my-assets-services",
		"version": "1.0.0",
		"license": {
			"name": "GPL-3.0"
		},
		"contact": {
			"name": "Jim Kim ",
			"email": "hjman.kim@gmail.com",
			"url": "http://https://my-assets-runner-rypg66fkzq-uc.a.run.app"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/exchange-rates": {
			"get": {
				"operationId": "GetExchangeRates",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExchangeRateResponseModel"
								}
							}
						}
					}
				},
				"description": "symbol: TICKER\r\nperiod: \"d\", // 'd' (daily), 'w' (weekly), 'm' (monthly), 'v' (dividends only)",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "symbol",
						"required": false,
						"schema": {
							"default": "DX-Y.NYB",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"default": "2021-09-08T07:18:46.943Z",
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"default": "2021-09-08T07:18:46.943Z",
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "period",
						"required": false,
						"schema": {
							"default": "d",
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}